+++ Changelog.md
@@ -2,6 +2,7 @@
 
 #1.1.0
 - Support des jets de soins
+- Fix stupéfié
 
 #1.0.0
 - Support des nouveaux formats de dégâts