+++ Changelog.md
@@ -1,5 +1,10 @@
 #Changelog
 
+#1.1.1
+- Gestion aveugle/aveuglé
+- Ajout Crocheter, Solliciter, Se faufiler
+- Fix Pris au dépourvu
+
 #1.1.0
 - Support des jets de soins
 - Fix stupéfié